move UITab code into corehq.tabs
