+++ corehq/tabs/__init__.py
@@ -1,0 +1,3 @@
+from .config import MENU_TABS
+
+__all__ = ['MENU_TABS']
