+++ corehq/tabs/config.py
@@ -1,16 +1,7 @@
-from corehq.apps.hqwebapp.models import (
-    DashboardTab,
-    ProjectInfoTab,
-    ReportsTab,
-    ProjectDataTab,
-    SetupTab,
-    ProjectUsersTab,
-    ApplicationsTab,
-    CloudcareTab,
-    MessagingTab,
-    AdminTab
-)
 from corehq.apps.styleguide.tabs import SGExampleTab
+from corehq.tabs.tabclasses import DashboardTab, ReportsTab, ProjectInfoTab, SetupTab, \
+    ProjectDataTab, ApplicationsTab, CloudcareTab, MessagingTab, ProjectUsersTab, \
+    AdminTab
 
 MENU_TABS = (
     DashboardTab,
